+++ python/bullmq/utils.py
@@ -1,3 +1,4 @@
+from typing import Any
 import semver
 import traceback
 import json
@@ -16,9 +17,25 @@ def extract_result(job_task, emit_callback):
             traceback.print_exc()
             emit_callback("error", e)
 
-def get_parent_key(opts: dict):
+def get_parent_key(opts: dict[str, str]):
     if opts:
         return f"{opts.get('queue')}:{opts.get('id')}"
 
 def parse_json_string_values(input_dict: dict[str, str]) -> dict[str, dict]:
     return {key: json.loads(value) for key, value in input_dict.items()}
+
+def object_to_flat_array(obj: dict[str, Any]) -> list[Any]:
+    """
+    Converts a dictionary into a flat array where each key is followed by its value.
+
+    Args:
+        obj (dict[str, Any]): The input dictionary to flatten.
+
+    Returns:
+        list[Any]: A flat list containing keys and values from the dictionary in order.
+    """
+    arr = []
+    for key, value in obj.items():
+        arr.append(key)
+        arr.append(value)
+    return arr